refractoring pack calculations
